Revert package renaming for git diff
