+++ hcl/hcl.go
@@ -2,7 +2,7 @@
  * Copyright SeatGeek
  * Licensed under the terms of the Apache-2.0 license. See LICENSE file in project root for terms.
  */
-package hclmerge
+package hcl
 
 import (
 	"fmt"
